refeator: 重构过大的MainWindow.swift，按照功能拆分到各个文件。
